@@ -2,9 +2,7 @@
 name = "rusty-slack-weather-status"
 version = "0.1.0"
 authors = ["shinespark <shnsprk@gmail.com>"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+edition = "2024"
 
 [dependencies]
 clap = "~2.27.0"
@@ -12,7 +10,7 @@ env_logger = "0.8.1"
 lazy_static = "1.4.0"
 log = "0.4"
 reqwest = { version = "0.10", features = ["json"] }
-rust-embed="5.6.0"
+rust-embed = "5.6.0"
 scraper = "0.12.0"
 tokio = { version = "0.2", features = ["full"] }
 toml = "0.5"
