@@ -3,7 +3,7 @@ extern crate clap;
 extern crate log;
 
 use clap::{App, Arg};
-use rusty_slack_weather_status::forecast::TenkiJpForecast;
+use rusty_slack_weather_status::models::tenki_jp_forecast::TenkiJpForecast;
 use rusty_slack_weather_status::slack::SlackRequest;
 
 #[tokio::main]
@@ -44,7 +44,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let is_dry_run = matches.is_present("DRY RUN");
     match is_dry_run {
         true => {
-            println!("{:?}", forecast);
+            println!("{forecast:?}");
             println!(
                 "{:?}, {:?}",
                 &forecast.build_emoji(),
