@@ -1,159 +1,19 @@
-use std::fmt::{self, Display, Formatter};
-use std::path::Path;
-
-use reqwest::StatusCode;
-use scraper::{Html, Selector};
-
 use crate::embed::{ALERT_EMOJI_MAP, WEATHER_EMOJI_MAP};
-
-const TRIM_CHARS: [char; 3] = ['[', '+', ']'];
-
-#[derive(Debug)]
-pub struct TenkiJpForecast {
-    status: StatusCode,
-    html: Html,
-}
+use crate::models::temp_diff::TempDiff;
 
 #[derive(Debug)]
 pub struct Forecast {
-    place: String,
-    date_time: String,
-    special_warnings: Option<Vec<String>>, // 特別警報
-    warnings: Option<Vec<String>>,         // 警報
-    alerts: Option<Vec<String>>,           // 注意報
-    weather: String,
-    weather_icon_name: String,
-    high_temp: i16,
-    high_temp_diff: TempDiff,
-    low_temp: i16,
-    low_temp_diff: TempDiff,
-}
-
-#[derive(Debug)]
-struct TempDiff {
-    temp_diff: i16,
-}
-
-impl Display for TempDiff {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self.temp_diff.is_positive() {
-            true => write!(f, "+{}", self.temp_diff),
-            false => write!(f, "{}", self.temp_diff),
-        }
-    }
-}
-
-impl TempDiff {
-    pub fn new(temp_diff: &str) -> Self {
-        Self {
-            temp_diff: temp_diff.parse::<i16>().unwrap(),
-        }
-    }
-}
-
-impl TenkiJpForecast {
-    pub async fn get(url: &str) -> Result<Self, Box<dyn std::error::Error>> {
-        let res = reqwest::get(url).await?;
-        Ok(Self {
-            status: res.status(),
-            html: Html::parse_document(&res.text().await?),
-        })
-    }
-
-    pub fn parse(&self) -> Result<Forecast, Box<dyn std::error::Error>> {
-        Ok(Forecast {
-            place: self.get_text("h2").split("の天気").collect::<Vec<_>>()[0].to_string(),
-            date_time: self
-                .get_text(".date-time")
-                .split("発表")
-                .collect::<Vec<_>>()[0]
-                .to_string(),
-            special_warnings: self.get_texts(".special-warn-entry"),
-            warnings: self.get_texts(".warn-entry"),
-            alerts: self.get_texts(".alert-entry"),
-            weather: self.get_text(".weather-telop"),
-            weather_icon_name: self.get_weather_icon_name(".weather-icon > img ", "src"),
-            high_temp: self
-                .get_text("dd.high-temp > .value")
-                .parse::<i16>()
-                .unwrap(),
-            high_temp_diff: TempDiff::new(
-                &self
-                    .get_text("dd.high-temp.tempdiff")
-                    .replace(TRIM_CHARS.as_ref(), ""),
-            ),
-            low_temp: self
-                .get_text("dd.low-temp > .value")
-                .parse::<i16>()
-                .unwrap(),
-            low_temp_diff: TempDiff::new(
-                &self
-                    .get_text("dd.low-temp.tempdiff")
-                    .replace(TRIM_CHARS.as_ref(), ""),
-            ),
-        })
-    }
-
-    fn get_text(&self, selector: &str) -> String {
-        let selector = Selector::parse(selector).unwrap();
-        match self.html.select(&selector).next() {
-            Some(x) => x
-                .text()
-                .collect::<Vec<_>>()
-                .into_iter()
-                .collect::<String>()
-                .trim()
-                .into(),
-            None => "".to_string(),
-        }
-    }
-
-    fn get_texts(&self, selector: &str) -> Option<Vec<String>> {
-        let selector = Selector::parse(selector).unwrap();
-        let texts = self
-            .html
-            .select(&selector)
-            .map(|x| {
-                x.text()
-                    .collect::<Vec<_>>()
-                    .into_iter()
-                    .collect::<String>()
-                    .trim()
-                    .into()
-            })
-            .collect::<Vec<_>>();
-
-        match texts.len() {
-            0 => None,
-            _ => Some(texts),
-        }
-    }
-
-    fn get_weather_icon_name(&self, selector: &str, attr: &str) -> String {
-        let weather_icon_path = self.get_attr(selector, attr).unwrap_or_default();
-        Self::get_file_stem(&weather_icon_path)
-    }
-
-    fn get_attr(&self, selector: &str, attr: &str) -> Option<String> {
-        let selector = Selector::parse(selector).unwrap();
-
-        self.html
-            .select(&selector)
-            .next()
-            .unwrap()
-            .value()
-            .attr(attr)
-            .map(|x| x.into())
-    }
-
-    fn get_file_stem(path: &str) -> String {
-        Path::new(path)
-            .file_stem()
-            .unwrap()
-            .to_str()
-            .unwrap()
-            .to_string()
-    }
+    pub place: String,
+    pub date_time: String,
+    pub special_warnings: Option<Vec<String>>, // 特別警報
+    pub warnings: Option<Vec<String>>,         // 警報
+    pub alerts: Option<Vec<String>>,           // 注意報
+    pub weather: String,
+    pub weather_icon_name: String,
+    pub high_temp: i16,
+    pub high_temp_diff: TempDiff,
+    pub low_temp: i16,
+    pub low_temp_diff: TempDiff,
 }
 
 impl Forecast {
@@ -200,28 +60,28 @@ impl Forecast {
             true => {
                 let special_warnings = self.special_warnings.as_ref().map(|x| {
                     x.iter()
-                        .map(|special_warning| format!("{}特別警報", special_warning))
+                        .map(|special_warning| format!("{special_warning}特別警報"))
                         .collect::<Vec<_>>()
                         .join(",")
                 });
 
                 let warnings = self.warnings.as_ref().map(|x| {
                     x.iter()
-                        .map(|warning| format!("{}警報", warning))
+                        .map(|warning| format!("{warning}警報"))
                         .collect::<Vec<_>>()
                         .join(",")
                 });
 
                 let alerts = self.alerts.as_ref().map(|x| {
                     x.iter()
-                        .map(|alert| format!("{}注意報", alert))
+                        .map(|alert| format!("{alert}注意報"))
                         .collect::<Vec<_>>()
                         .join(",")
                 });
 
                 let texts = vec![special_warnings, warnings, alerts]
                     .into_iter()
-                    .filter_map(|x| x)
+                    .flatten()
                     .collect::<Vec<_>>()
                     .join(",");
 
@@ -255,85 +115,6 @@ impl Forecast {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn test_get_text() {
-        let tenki_jp_forecast = TenkiJpForecast {
-            status: Default::default(),
-            html: Html::parse_document(
-                "<html><h1>h1要素</h1><h2>h2要素</h2><h3>h3要素</h3></html>",
-            ),
-        };
-
-        assert_eq!(tenki_jp_forecast.get_text("h2"), "h2要素");
-    }
-
-    mod test_get_texts {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        #[test]
-        fn found() {
-            let tenki_jp_forecast = TenkiJpForecast {
-                status: Default::default(),
-                html: Html::parse_document(
-                    "<html><span class='alert-entry'>洪水</span><span class='alert-entry'>雷</span></html>",
-                ),
-            };
-
-            assert_eq!(
-                tenki_jp_forecast.get_texts(".alert-entry"),
-                Some(vec!["洪水".to_string(), "雷".to_string()])
-            );
-        }
-
-        #[test]
-        fn not_found() {
-            let tenki_jp_forecast = TenkiJpForecast {
-                status: Default::default(),
-                html: Html::parse_document("<html></html>"),
-            };
-
-            assert_eq!(tenki_jp_forecast.get_texts(".alert-entry"), None);
-        }
-    }
-
-    #[test]
-    fn test_get_attr() {
-        let tenki_jp_forecast = TenkiJpForecast {
-            status: Default::default(),
-            html: Html::parse_document(
-                "<html><img src='https://static.tenki.jp/images/icon/forecast-days-weather/12.png'></html>",
-            ),
-        };
-
-        assert_eq!(
-            tenki_jp_forecast.get_attr("img", "src"),
-            Some("https://static.tenki.jp/images/icon/forecast-days-weather/12.png".to_string())
-        );
-    }
-
-    mod test_get_file_stem {
-        use super::*;
-        use pretty_assertions::assert_eq;
-
-        #[test]
-        fn daytime() {
-            let str = "https://static.tenki.jp/images/icon/forecast-days-weather/12.png";
-            let result = TenkiJpForecast::get_file_stem(str);
-
-            assert_eq!(result, "12".to_string());
-        }
-
-        #[test]
-        fn night() {
-            let str = "https://static.tenki.jp/images/icon/forecast-days-weather/12_n.png";
-            let result = TenkiJpForecast::get_file_stem(str);
-
-            assert_eq!(result, "12_n".to_string());
-        }
-    }
 
     mod test_build_emoji {
         use super::*;
